Soft keyboard Close Work around
I used an extension to add a done button above the soft keyboard. when pressed this closes the softkeyboard.  from some reason the typical delegate method through UITextViewDelegate is not working.. can come back to this but it is functional now.

Also rearranged the layout so that the textview is higher so that the softkeyboard doesn't cover it on a small screen.

Next...
1. Fix this workaround (above)
2. Make it pretty